servidor/requirements.txt
Flask
matplotlib
numpy
pyserial
